% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybrr-fit.R
\name{phybrr}
\alias{phybrr}
\alias{phybrr.default}
\alias{phybrr.data.frame}
\alias{phybrr.matrix}
\alias{phybrr.formula}
\alias{phybrr.recipe}
\title{Fit a \code{phybrr}}
\usage{
phybrr(x, ...)

\method{phybrr}{default}(x, ...)

\method{phybrr}{data.frame}(x, y, ...)

\method{phybrr}{matrix}(x, y, ...)

\method{phybrr}{formula}(formula, data, ...)

\method{phybrr}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{phybrr} object.
}
\description{
\code{phybrr()} fits a model.
}
\examples{
predictors <- mtcars[, -1]
outcome <- mtcars[, 1]

# XY interface
mod <- phybrr(predictors, outcome)

# Formula interface
mod2 <- phybrr(mpg ~ ., mtcars)

# Recipes interface
library(recipes)
rec <- recipe(mpg ~ ., mtcars)
rec <- step_log(rec, disp)
mod3 <- phybrr(rec, mtcars)

}
